remove files from config and update with go config
This removes a bunch of files that I don't want tracked by this repo.
They never should have been in there to begin with, but I made a
mistake. The other changes center around fixing autocmds by using groups
and correctly unsetting them to prevent duplicates. There is some config
included for nvim with using go and vim-go as well.